+++ src/NimBleClient.h
@@ -6,9 +6,11 @@
 
 namespace sensirion::upt::ble_auto_detection{
 
+static constexpr uint32_t SCAN_DURATION_MS = 30 * 1000; // 30 seconds scans.
+
 class NimBleClient
     : public BleClient,
-      public NimBLEAdvertisedDeviceCallbacks {
+      public NimBLEScanCallbacks {
   public:
     NimBleClient();
     ~NimBleClient();
@@ -21,7 +23,8 @@ class NimBleClient
     BleClientCallback* _callback;
     void setupBleScans();
     void startBleScans();
-    void onResult(NimBLEAdvertisedDevice* advertisedDevice) override;
+    void onResult(const NimBLEAdvertisedDevice* advertisedDevice) override;
+    void onScanEnd(const NimBLEScanResults& results, int reason) override;
 };
 } // end namespace sensirion::upt::ble_auto_detection
 #endif /* SENSIRION_UPT_BLE_AUTO_DETECTION_NIMBLE_CLIENT_H */